Moved all tests into their own file to make outer-testset easier.
